# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  types.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  types.LoginResponse:
    properties:
      token:
        type: string
    type: object
  types.LoginUserPayload:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  types.RegisterUserPayload:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        maxLength: 130
        minLength: 3
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    type: object
...
  types.UpdatePasswordPayload:
    properties:
      currentPassword:
        maxLength: 130
        minLength: 3
        type: string
      newPassword:
        maxLength: 130
        minLength: 3
        type: string
    required:
    - currentPassword
    - newPassword
    type: object
  types.UpdateProfilePayload:
    properties:
      firstName:
        maxLength: 255
        minLength: 1
        type: string
      lastName:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - firstName
    - lastName
    type: object
...
  types.UserLookup:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  types.UserProfile:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for task tracking with goals and assignments
  title: Task Tracker API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.LoginUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Login
      tags:
      - auth
  /profile:
    get:
      description: Get authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserProfile'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update authenticated user's first and last name
      parameters:
      - description: Profile payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.UpdateProfilePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - users
  /profile/password:
    put:
      consumes:
      - application/json
      description: Change authenticated user's password
      parameters:
      - description: Password payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.UpdatePasswordPayload'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update password
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.RegisterUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Register
      tags:
      - auth
...
  /users/lookup:
    get:
      description: List users for assignment lookups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.UserLookup'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users lookup
      tags:
      - users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"